feat: Show unsupported network error in snackbar instead of pair screen
